fix: use bigint for size columns

+++ internal/migrations/sql/25_artifact_size.up.sql
@@ -1,4 +1,4 @@
-ALTER TABLE ArtifactVersion ADD COLUMN manifest_blob_size INTEGER NOT NULL DEFAULT 0;
-ALTER TABLE ArtifactVersionPart ADD COLUMN artifact_blob_size INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE ArtifactVersion ADD COLUMN manifest_blob_size BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE ArtifactVersionPart ADD COLUMN artifact_blob_size BIGINT NOT NULL DEFAULT 0;
 ALTER TABLE ArtifactVersion ALTER COLUMN manifest_blob_size DROP DEFAULT;
 ALTER TABLE ArtifactVersionPart ALTER COLUMN artifact_blob_size DROP DEFAULT;
